Update for the variance scaling parameter.

Update for the variance scaling parameter.

back_update_xi(betahat_ols, S_diag, Amat, mubeta, mubeta_matrix,
  sig2beta_matrix, gamma_mat, muv, Sigma_v, phi, var_inflate_pen = 0)

Arguments

betahat_ols
A vector of numerics. The oridinary least squares estimates of the regression coefficients.

S_diag
A vector of positive numerics. The standard errors of betahat_ols.

Amat
The A matrix for the variational EM.

mubeta
The current means of the betas

mubeta_matrix
The current mixing means of the variational densities of the betas.

sig2beta_matrix
The current mixing variances of the variational densities of the betas.

gamma_mat
The current mixing proportions of the variational densities of the betas.

muv
A matrix of one column of numerics. The current mean of v.

Sigma_v
The current covariance matrix of the latent v.

phi
The current "g" hyperparameter.

var_inflate_pen
The penalty to apply on the variance inflation parameter. Defaults to 0, but should be something non-zero when alpha = 1 and scale_var = TRUE.