Second step of the backwash procedure.

Second step of the backwash procedure.

backwash_second_step(betahat_ols, S_diag, alpha_tilde, tau2_seq, lambda_seq,
  pi_init_type = c("zero_conc", "uniform", "random"), scale_var = TRUE,
  sprop = 0, var_inflate_pen = 0, verbose = TRUE)

Arguments

betahat_ols
A vector of numerics. The oridinary least squares estimates of the regression coefficients.

S_diag
A vector of positive numerics. The standard errors of betahat_ols.

alpha_tilde
A matrix of numerics. The estimated coefficients of the confounders.

tau2_seq
A vector of positive numerics. The known grid of prior mixing variances.

lambda_seq
A vector of penalties for the estimate of the mixing proportions.

pi_init_type
How should we initialize the mixing proportions? By concentrating on zero ("zero_conc"), by equal weights on all mixing distributions ("uniform"), or by sampling uniformly on the simplex ("random")?

scale_var
A logical. Should we estimate a variance inflation parameter (TRUE) or not (FALSE)?

sprop
If \(b\) is an effect and \(s\) is an estimated standard error, then we model \(b/s^{sprop}\) as exchangeable. The default is 0. When sprop = 1, for identifiability reasons it must be the case that scale_var = FALSE.

var_inflate_pen
The penalty to apply on the variance inflation parameter. Defaults to 0, but should be something non-zero when alpha = 1 and scale_var = TRUE.

verbose
If verbose = TRUE, print progress of the algorithm to the console.