Test whether a user-specified function is compatible with ruvb.

This function tests whether a user-defined function fa_func is compatible for use in ruvb. Let n be the number of samples, p be the number of genes, m be the number of control genes, and q be the number of covariates. In general, fa_func needs to take as input Y21 a q by m matrix of numerics, Y31 a (n - q) by m matrix of numerics, Y32 a (n - q) by (p - m) matrix of numerics, and k a positive integer that is the rank of the factor model if assuming a factor model. fa_func also needs to return Y22_array a three-way array whose first two dimensions are k and (p - m).

fa_tester_ruvb(fa_func, fa_args = list())

Arguments

fa_func
The factor analysis function.

fa_args
Additional arguments to pass to fa_func.

Value

A list with some or all of the following elements:

ok A logical. A value of TRUE means the function is OK. A value of FALSE means the function is not OK.

why A string vector giving the reason(s) why the function is not OK if ok = FALSE.

input A list of input values to help you debug.

See also

Author

David Gerard