A wrapper for fix_caw so that I can use SQUAREM.

A wrapper for fix_caw so that I can use SQUAREM.

fix_caw_wrapper(wzs_vec, rotate_out, weight_func = ash_wrap,
  weight_args = list(), min_scale = 0.8)

Arguments

wzs_vec
A vector of numerics. The first p of which are the w_current, the next length(cov_of_interest) * k of which are Z2hat, the last of which is scale_val.

rotate_out
The output from rotate_model with degrees_freedom added to the list.

weight_func
The function that calculates the weights given w_current. So far, this only supports functions that take as input three quantities: betahat, sebetahat, and degrees_freedom

weight_args
Additional arguments to pass to weight_func.

min_scale
A scalar. The minimum value that scale_val is allowed to take. In practice, you wouldn't think that shrinking the variance is a good idea. But a value of 1 seems to stringent. So I set this to 0.8 to give you some wiggle room. Most of the time scale_val is usually estimated greater than 1.