Quasi-mle when first vr rows of Y have variances which differ by a scale factor from the rest of the rows.

This one will iterate through a grid of lambdas and choose the one with the highest likelihood.

qmle_ruv2_lambda_grid(Y, r, vr, gridsize = 20, gridlow = 1, gridhigh = 10,
  tol = 10^-3, maxit = 1000, plot_llike = FALSE)

Arguments

Y
A matrix of numerics. The data.

r
the rank.

vr
The number of the first few rows whose variances differ by a multiplicative factor.

gridsize
A positive integer. The size of the grid for the variance inflation parameter.

gridlow
A positive numeric. The minimum variance inflation parameter to try out.

gridhigh
A positive numeric. The maximum variance inflation parameter to try out.

tol
The function tolerance for stopping. A positive numeric.

maxit
The maximum number of iterations to run.

plot_llike
A logical. Should I plot the profiled log-likelihood?

Value

alpha A matrix of the estimated factor loadings.

Z A matrix of the estimated factors.

sig_diag A vector of the estimated column-wise variances.

lambda A numeric scalar of the estimated variance inflation parameter.