Very inefficient copy of Rubin and Thayer iteration mostly meant for debugging.

Very inefficient copy of Rubin and Thayer iteration mostly meant for debugging.

rubin_copy(alpha_sigma, S, n, p, k)

Arguments

alpha_sigma
A vector. The first k * p elements are the vectorization of alpha. The last elements are sig_diag.

S
A matrix. This is the sample covariance matrix of the samples that don't have a missing block.

n
The sample size

p
The number of genes.

k
The rank of the mean.

Author

David Gerard