Wrapper for tregress_obj.

This is mostly so I can run stats::optim with just lambda.

tregress_obj_wrapper(lambda, Z, Y, alpha, sig_diag, nu)

Arguments

lambda
A positive numeric. The variance inflation parameter.

Z
A matrix of numerics with one column. The coefficients.

Y
A matrix of numerics with one column. The response variables.

alpha
A matrix of numerics, the covariates. It must be that nrow(Y) is equal to nrow(alpha).

sig_diag
A vector of numerics. The variances of the elements in Y, but only assumed to be known up to a scaling factor.

nu
A positive numeric scalar. The known degrees of freedom of the t-distribution.