Log-likelihood when using t errors and mixture of uniforms prior.

Log-likelihood when using t errors and mixture of uniforms prior.

uniform_mix_llike(pi_vals, z2, xi, betahat_ols, S_diag, alpha_tilde, a_seq,
  b_seq, lambda_seq, degrees_freedom, var_inflate_pen = 0)

Arguments

pi_vals
A vector of non-negative numerics that sums to one. The current values of the mixing proportions.

z2
A vector of numerics. The current values of the unobserved confounders.

xi
A positive numeric. The current value of the variance inflation factor.

betahat_ols
A vector of numerics. The OLS regression coefficients.

S_diag
A vector of positive numerics. The standard errors.

alpha_tilde
A matrix of numerics. The estimats of the coefficients of the confounders.

a_seq
A vector of numerics (usually non-positive). The lower bounds of the uniform mixing densities.

b_seq
A vector of numerics (usually non-negative). The upper bounds of the uniform mixing densities.

lambda_seq
A vector of numerics greater than 1. The penalties on the mixing proportions.

degrees_freedom
A positive scalar or a vector of positive scalars the length of betahat_ols. The degrees of freedom of the t-distribution.

var_inflate_pen
The penalty to apply on the variance inflation parameter. Defaults to 0, but should be something non-zero when alpha = 1 and scale_var = TRUE.