vicar: Various Ideas for Confounder Adjustment in Regression.

vicar contains functions for implementing RUV3, a Bayesian version of RUV, and a calibrated version of RUV4. It also contains the functions mouthwash and backwash for confounder adjustment when one does not have access to control genes.

Main <code>vicar</code> functions

ruv3: An implementation of RUV3.

ruvb: An implementation of RUV-Bayes.

vruv4: Variance inflated version of RUV4/CATE.

mouthwash: Empirical Bayesian confounder adjustment without control genes.

backwash: Variational Bayesian confounder adjustment without control genes.

Vignettes (run in R)

utils::vignette("customFA", package = "vicar") For providing custom factor analyses during confounder adjustment.

utils::vignette("custom_prior", package = "vicar") For providing custom priors in ruvb.

utils::vignette("sample_analysis", package = "vicar") For a sample analysis on a simulated dataset using the vicar functions and other confounder adjustment packages.

If the vignette code does not work, then you probably did not build them during install. See https://github.com/dcgerard/vicar#vignettes.