QR rotation to independent models.

QR rotation to independent models.

rotate_model(Y, X, k, cov_of_interest = ncol(X),
  include_intercept = TRUE, limmashrink = FALSE, fa_func = pca_naive,
  fa_args = list(), do_factor = TRUE)

Arguments

Y
A matrix of numerics. These are the response variables where each column has its own variance. In a gene expression study, the rows are the individuals and the columns are the genes.

X
A matrix of numerics. The covariates of interest.

k
A non-negative integer.The number of unobserved confounders. If not specified and the R package sva is installed, then this function will estimate the number of hidden confounders using the methods of Buja and Eyuboglu (1992).

cov_of_interest
A vector of positive integers. The column numbers of the covariates in X whose coefficients you are interested in. The rest are considered nuisance parameters and are regressed out by OLS.

include_intercept
A logical. If TRUE, then it will check X to see if it has an intercept term. If not, then it will add an intercept term. If FALSE, then X will be unchanged.

limmashrink
A logical. Should we apply hierarchical shrinkage to the variances (TRUE) or not (FALSE)? If degrees_freedom = NULL and limmashrink = TRUE and likelihood = "t", then we'll also use the limma returned degrees of freedom.

fa_func
A factor analysis function. The function must have as inputs a numeric matrix Y and a rank (numeric scalar) r. It must output numeric matrices alpha and Z and a numeric vector sig_diag. alpha is the estimate of the coefficients of the unobserved confounders, so it must be an r by ncol(Y) matrix. Z must be an r by nrow(Y) matrix. sig_diag is the estimate of the column-wise variances so it must be of length ncol(Y). The default is the function pca_naive that just uses the first r singular vectors as the estimate of alpha. The estimated variances are just the column-wise mean square.

fa_args
A list. Additional arguments you want to pass to fa_func.

do_factor
A logical. Should we do the factor analysis or just rotation?

Value

A list that contains some of the following elements.

betahat_ols: The OLS estimates of the coefficients of interest.

Q: The Q from the QR decomposition of X

R11: The top left element of the R matrix from the QR decomposition of X

R12: The top right element of the R matrix from the QR decomposition of X

R22: The bottom right element of the R matrix from the QR decomposition of X

Y1: The rows of the rotated Y matrix that correspond to the uninteresting covariates in X

Y2: The rows of the rotated Y matrix that correspond to the interesting covariates in X

Y3: The rows of the rotated Y matrix that correspond to the nullspace of X

k: The number of latent factors. Estimated with num.sv if not provided in the function arguments.

X: The provided X matrix. This might differ from the original X matrix up to a permutation of the columns. The interesting covariates are placed at the end of X and the uninteresting covariates are placed at the start of X.

Y_tilde: The rotated Y matrix.

alpha: The estimates of the coefficients of the unobserved confounders. Only returned if do_factor = TRUE

Z3: The estimates of the unobserved confounders (third part). Only returned if do_factor = TRUE

sig_diag: The estimates of the column-specific variances (NOT standard deviations). A vector. Only returned if do_factor = TRUE

prior_df: The estimated prior degrees of freedom. Returns NULL if limmashrink = FALSE. Only returned if do_factor = TRUE