A simulated RNA-seq dataset.

This is a simulated dataset that shares characteristics with the GTEx data.

Format

A list with the following elements

Y

A matrix of gene expression levels whose rows index the samples and whose columns index the genes.

X

A matrix of covariates. The first column is a vector of ones for an intercept term and the second column is a "treatment" indicator.

ctl

A logical vector where a TRUE at location \(i\) indicates the prensence of a control gene at location \(i\) in Y and a FALSE at location \(i\) indicates the absence of a control gene at location \(i\) of Y.

beta

The two effect sizes for treatment vs control.

which_null

A logical vector where a TRUE indicates the location of a null gene.

Source

If you want the real GTEx data, go to http://www.gtexportal.org/home/.